Made app configurable without modifying any files
Settings are loaded from the file pointed to by the SETTINGS_PATH
environment variable. That file contains the location of all settings,
including a path for the database configuration and for logging.

This allows the executor of this application to customize all settings
without creating or modifying any files within the application.