Check database existence before attempting delete
`DROP DATABASE` returns the number of *tables* deleted, not the number
of databases. The only way to see if a database was (likely) deleted is
by looking it up with a separate query first.

This isn't perfect, as even if another request beat us to the actual
delete, both requests would return a 204. But we don't anticipate any
negative side effects from such a scenario.

[#58124456]